:recycle: Move util function to common
